Character outfit handler finished
The outfit handler of the character is done. It has been tested with a custom selector.
Tested using 4 types of outfits:
- Bottom (Underwears and boxers)
- Top: (Full-body outfits)
- Hair
- Hats

Every outfit is a scriptable object that contains a sprite, and a prefab to instantiate in the player. Hats have a property that overrides the hair in case the combination of hair and hat is not good.